➡️ Move inventory API classes to their own folder
